+++ BunqSdk/Model/Core/BunqModel.cs
@@ -39,15 +39,15 @@ namespace Bunq.Sdk.Model.Core
         }
 
         /// <summary>
-        /// De-serializes an ID object and returns its integer value.
+        /// De-serializes an ID object and returns its long value.
         /// </summary>
-        protected static BunqResponse<int> ProcessForId(BunqResponseRaw responseRaw)
+        protected static BunqResponse<long> ProcessForId(BunqResponseRaw responseRaw)
         {
             var responseItemObject = GetResponseItemObject(responseRaw);
             var unwrappedItemJsonString = GetUnwrappedItemJsonString(responseItemObject, FIELD_ID);
             var responseValue = BunqJsonConvert.DeserializeObject<Id>(unwrappedItemJsonString).IdInt;
 
-            return new BunqResponse<int>(responseValue, responseRaw.Headers);
+            return new BunqResponse<long>(responseValue, responseRaw.Headers);
         }
 
         private static JObject GetResponseItemObject(BunqResponseRaw responseRaw)
